@@ -1,45 +1,40 @@
 {
-  "name": "comp2156",
-  "version": "1.0.0",
-  "description": "Hello World",
-  "main": "index.js",
-  "scripts": {
-    "start": "react-scripts --openssl-legacy-provider start",
-    "build": "react-scripts --openssl-legacy-provider build",
-    "predeploy": "yarn run build",
-    "deploy": "gh-pages -d build",
-    "test": "echo \"Warning: No tests created yet.\"",
-    "eject": "react-scripts eject"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/joshua-ensou/comp2156.git"
-  },
-  "keywords": [],
-  "author": "Joshua Marcelo",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/joshua-ensou/comp2156/issues"
-  },
-  "homepage": "https://github.com/joshua-ensou/comp2156#readme",
+  "name": "devops_comp2156_lab6",
+  "version": "0.1.0",
+  "private": true,
   "dependencies": {
-    "gh-pages": "^2.0.1",
-    "react": "^16.8.3",
-    "react-dom": "^16.8.3",
-    "react-scripts": "2.1.5",
-    "travis": "^0.1.1",
-    "yarn": "^1.13.0"
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "5.0.1",
+    "web-vitals": "^2.1.4"
+  },
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "predeploy": "npm run build",
+    "test": "echo \"Warning: No tests created yet.\"",
+    "eject": "react-scripts eject",
+    "deploy": "gh-pages -d build"
   },
   "eslintConfig": {
-    "extends": "react-app"
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
   },
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not ie <= 11",
-    "not op_mini all"
-  ],
-  "devDependencies": {
-    "webpack": "^4.28.3"
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
