@@ -1,25 +1,11 @@
-import React, { Component } from 'react';
 import './App.css';
-import Sidebar from './components/sidebar'
-import Introduction from './components/introduction'
-import About from './components/about'
-import Timeline from './components/timeline'
 
-class App extends Component {
-  render() {
-    return (
-      <div id="colorlib-page">
-        <div id="container-wrap">
-         	<Sidebar></Sidebar>
-          <div id="colorlib-main">
-            <Introduction></Introduction>
-            <About></About>
-            <Timeline></Timeline>
-          </div>
-      	</div>
-      </div>
-    );
-  }
+function App() {
+  return (
+    <div>
+      <h1>Hello DevOps</h1>
+    </div>
+  );
 }
 
 export default App;
